pyproject.toml
[project]
name = "test-redis-fulltext-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "redis",
    "redisearch",
    "azure-search-documents",
    "python-dotenv",
]
